@@ -25,8 +25,8 @@
     "match-sorter": "^6.3.0",
     "msw": "^0.29.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-error-boundary": "^3.1.3",
     "react-icons": "^4.2.0",
     "react-query": "2.1.1",
@@ -37,13 +37,12 @@
     "stop-runaway-react-effects": "^2.0.0"
   },
   "devDependencies": {
-    "@testing-library/cypress": "^7.0.6",
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/react-hooks": "^7.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
+    "@testing-library/cypress": "^8.0.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.0",
+    "@testing-library/user-event": "^14.1.0",
+    "@types/react": "^18.0.4",
+    "@types/react-dom": "^18.0.0",
     "cross-env": "^7.0.3",
     "cypress": "^7.5.0",
     "cypress-hmr-restarter": "^2.0.2",
