@@ -16,6 +16,10 @@ import {App} from 'app'
 
 const apiURL = process.env.REACT_APP_API_URL
 
+const fakeTimerUserEvent = userEvent.setup({
+  advanceTimers: () => jest.runOnlyPendingTimers(),
+})
+
 async function renderBookScreen({user, book, listItem} = {}) {
   if (user === undefined) {
     user = await loginAsUser()
@@ -71,7 +75,7 @@ test('can create a list item for the book', async () => {
   await renderBookScreen({listItem: null})
 
   const addToListButton = screen.getByRole('button', {name: /add to list/i})
-  userEvent.click(addToListButton)
+  await userEvent.click(addToListButton)
   expect(addToListButton).toBeDisabled()
 
   await waitForLoadingToFinish()
@@ -102,7 +106,7 @@ test('can remove a list item for the book', async () => {
   const removeFromListButton = screen.getByRole('button', {
     name: /remove from list/i,
   })
-  userEvent.click(removeFromListButton)
+  await userEvent.click(removeFromListButton)
   expect(removeFromListButton).toBeDisabled()
 
   await waitForLoadingToFinish()
@@ -134,7 +138,7 @@ test('can mark a list item as read', async () => {
   await renderBookScreen({user, book, listItem})
 
   const markAsReadButton = screen.getByRole('button', {name: /mark as read/i})
-  userEvent.click(markAsReadButton)
+  await userEvent.click(markAsReadButton)
   expect(markAsReadButton).toBeDisabled()
 
   await waitForLoadingToFinish()
@@ -161,8 +165,8 @@ test('can edit a note', async () => {
   const newNotes = faker.lorem.words()
   const notesTextarea = screen.getByRole('textbox', {name: /notes/i})
 
-  userEvent.clear(notesTextarea)
-  userEvent.type(notesTextarea, newNotes)
+  await fakeTimerUserEvent.clear(notesTextarea)
+  await fakeTimerUserEvent.type(notesTextarea, newNotes)
 
   // wait for the loading spinner to show up
   await screen.findByLabelText(/loading/i)
@@ -196,8 +200,8 @@ describe('console errors', () => {
   })
 
   test('note update failures are displayed', async () => {
-    // using fake timers to skip debounce time
     jest.useFakeTimers()
+    // using fake timers to skip debounce time
     await renderBookScreen()
 
     const newNotes = faker.lorem.words()
@@ -213,7 +217,7 @@ describe('console errors', () => {
       }),
     )
 
-    userEvent.type(notesTextarea, newNotes)
+    await fakeTimerUserEvent.type(notesTextarea, newNotes)
     // wait for the loading spinner to show up
     await screen.findByLabelText(/loading/i)
     // wait for the loading spinner to go away
