@@ -16,14 +16,14 @@ test('can be opened and closed', async () => {
       </ModalContents>
     </Modal>,
   )
-  userEvent.click(screen.getByRole('button', {name: 'Open'}))
+  await userEvent.click(screen.getByRole('button', {name: 'Open'}))
 
   const modal = screen.getByRole('dialog')
   expect(modal).toHaveAttribute('aria-label', label)
   const inModal = within(screen.getByRole('dialog'))
   expect(inModal.getByRole('heading', {name: title})).toBeInTheDocument()
 
-  userEvent.click(inModal.getByRole('button', {name: /close/i}))
+  await userEvent.click(inModal.getByRole('button', {name: /close/i}))
 
   expect(screen.queryByRole('dialog')).not.toBeInTheDocument()
 })
